Realized determination of element's material (Now the program works)

+++ Element.h
@@ -4,5 +4,5 @@ using namespace std;
 
 struct Element {
 	vector<int> nodesNumbers;
-	double k;
+	int k;
 };
